Hexen: remove BLOCKMAP limit, re-create if necessary

+++ src/hexen/p_local.h
@@ -271,8 +271,8 @@ void P_RadiusAttack(mobj_t * spot, mobj_t * source, int damage, int distance,
 // ***** P_SETUP *****
 
 extern byte *rejectmatrix;      // for fast sight rejection
-extern short *blockmaplump;     // offsets in blockmap are from here
-extern short *blockmap;
+extern int32_t *blockmaplump;     // offsets in blockmap are from here
+extern int32_t *blockmap;
 extern int bmapwidth, bmapheight;       // in mapblocks
 extern fixed_t bmaporgx, bmaporgy;      // origin of block map
 extern mobj_t **blocklinks;     // for thing chains
